git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@5962 0bcb941b-74d1-0310-bb26-e6636707f451

+++ mathfun/graphMatching/maxWeightedMatching.cpp
@@ -66,7 +66,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
   double* p = mxGetPr(prhs[1]);
   double* q = mxGetPr(prhs[2]);
 
-  typedef SmartGraph::EdgeMap<int> WeightMap;
+  typedef SmartGraph::EdgeMap<double> WeightMap;
 
   WeightMap weight_map(g);
